+++ src/composable/chat/useChatProperties.ts
@@ -8,11 +8,15 @@ const data = reactive({
 	useHighContrastColors: true,
 	showTimestamps: false,
 	showModerationIcons: false,
+	hovering: false,
+	pauseReason: new Set<ChatPauseReason>(["SCROLL"]),
 	currentChannel: {} as CurrentChannel,
 	imageFormat: "WEBP" as SevenTV.ImageFormat,
 	twitchBadgeSets: {} as Twitch.BadgeSets | null,
 	blockedUsers: new Set<string>(),
 });
+
+type ChatPauseReason = "MOUSEOVER" | "SCROLL" | "ALTKEY";
 
 export function useChatProperties() {
 	return data;